Move code to enclose.js
