+++ contrib/render-a11y-string/render-a11y-string.js
@@ -387,15 +387,6 @@ const handleObject = (
             break;
         }
 
-        case "phase": {
-            buildRegion(a11yStrings, function(a11yStrings) {
-                a11yStrings.push("start phase angle");
-                buildA11yStrings(tree.body, a11yStrings, atomType);
-                a11yStrings.push("end phase angle");
-            });
-            break;
-        }
-
         case "phantom": {
             a11yStrings.push("empty space");
             break;
@@ -541,6 +532,13 @@ const handleObject = (
                     regionStrings.push("start strikeout");
                     buildA11yStrings(tree.body, regionStrings, atomType);
                     regionStrings.push("end strikeout");
+                });
+                break;
+            } else if (/phase/.test(tree.label)) {
+                buildRegion(a11yStrings, function(regionStrings) {
+                    regionStrings.push("start phase angle");
+                    buildA11yStrings(tree.body, regionStrings, atomType);
+                    regionStrings.push("end phase angle");
                 });
                 break;
             }
